requirements.txt
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
Pillow==10.1.0
pyjwt==2.8.0
python-dotenv==1.0.0